pyproject.toml
[project]
name = "fastapitasks"
version = "1.0.0"
description = ""
authors = [
    {name = "damir",email = "damiraminov1@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.123.5,<0.124.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "argon2-cffi-bindings (>=25.1.0,<26.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "celery (>=5.6.0,<6.0.0)",
    "amqp (>=5.3.1,<6.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
